refactor trifold components structure, create nav style css
