--- src/components/groups-suggestions-sidebar/groups-suggestion-item.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import GroupType from "../../modules/groupType";
-
-const GroupSuggestionItem = (group: GroupType) => {
-    return (
-        <span>
-            <img className="rounded-circle" height={48} 
-                src={`${group.image}`}
-                alt={group.name}
-            />
-            <p> {group.name} </p>
-        </span>
-    );
-};
-
-export default GroupSuggestionItem;
